refactor: move testing module to kloppy._utils
