+++ kloppy/tests/test_metrica_epts.py
@@ -7,6 +7,7 @@ import pytest
 
 from kloppy import metrica
 from kloppy._utils import performance_logging
+from kloppy._utils.testing import skip_if_no
 from kloppy.domain import Orientation, Point, Provider, Score
 from kloppy.infra.serializers.tracking.metrica_epts.metadata import (
     _load_provider,
@@ -16,7 +17,6 @@ from kloppy.infra.serializers.tracking.metrica_epts.reader import (
     build_regex,
     read_raw_data,
 )
-from kloppy.tests.testing import skip_if_no
 
 
 class TestMetricaEPTSTracking:
